reorganize tests add test for train loss sanity
